third_party/blink/web_tests/fast/css/fontface-arraybuffer-expected.txt

CONSOLE WARNING: line 29: SharedArrayBuffer will require cross-origin isolation as of M91, around May 2021. See https://developer.chrome.com/blog/enabling-shared-array-buffer/ for more details.
Tests ArrayBuffer / ArrayBufferView constructors of FontFace.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS face1.status is "loaded"
PASS face2.status is "loaded"
PASS rejectionValue.name is "TypeError"
PASS face3.status is "error"
PASS rejectionValue.name is "SyntaxError"
PASS document.getElementById('FontFromArrayBuffer').offsetWidth is document.getElementById('ref').offsetWidth
PASS document.getElementById('FontFromArrayBufferView').offsetWidth is document.getElementById('ref').offsetWidth
PASS successfullyParsed is true

TEST COMPLETE

abc
abc
abc
